Did some reorganizing, made a sub directory named vuforia under vision
