+++ java/examples/VuforiaParticleDetectionExample.java
@@ -11,7 +11,7 @@ import team25core.ColorThiefTask;
 import team25core.GamepadTask;
 import team25core.Robot;
 import team25core.RobotEvent;
-import team25core.VuforiaBase;
+import team25core.vision.vuforia.VuforiaBase;
 
 @Autonomous(name = "VuforiaParticleDetection")
 @Disabled
