+++ java/team25core/TouchEnabledGLSurfaceView.java
@@ -1,19 +1,11 @@
 package team25core;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.Color;
-import android.opengl.GLException;
 import android.view.MotionEvent;
 
-import com.qualcomm.robotcore.util.RobotLog;
-
-import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.internal.opengl.AutoConfigGLSurfaceView;
 
-import java.nio.IntBuffer;
-
-import javax.microedition.khronos.opengles.GL10;
+import team25core.vision.vuforia.VuforiaLocalizerCustom;
 
 public class TouchEnabledGLSurfaceView extends AutoConfigGLSurfaceView {
