+++ java/team25core/vision/vuforia/VuforiaLCRTask.java
@@ -1,12 +1,15 @@
-package team25core;
-
-import com.qualcomm.robotcore.util.ElapsedTime;
+package team25core.vision.vuforia;
 
 import org.firstinspires.ftc.robotcore.external.ClassFactory;
 import org.firstinspires.ftc.robotcore.external.navigation.RelicRecoveryVuMark;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackable;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackables;
+
+import team25core.PeriodicTimerTask;
+import team25core.Robot;
+import team25core.RobotEvent;
+import team25core.RobotTask;
 
 /**
  * Created by admin on 10/21/2017.
